requirements.txt
torch>=2.8.0
torchvision>=0.23.0
einops==0.7.0
PyMuPDF
Pillow